WTF IS GOING ONNNNN!!!
I did all this work just for the javascript code to not work for some reason. I spent hours trying to figure out why but seeing how I am just a fairly new to js, I dont understand shit. But on the sunny side, I fixed my issue with the icons not going to the right.