Fixed the issue of my dropdown menu
So basically, the drop menu action only was initialized only if there was anything occupying the drop button. The problem occurs because my icons are not text, so the code reads as the div is empty rather than actually well not empty.

My roommate Jacob (IT major...lul) helped me identify the code in the javascript that read my div as empty.