The power of counter!
I solved my problem of updating the # of the playlist. I went for incorporating a counter and using replaceWith() method. I originally thought of using the index and somehow using that data to update each cloned playlist. There is probably a way to do that but I could not figure it out and so I figured I was just overcomplicating the solution to my problem. Another Victory!!!