FINALLY FIGURED OUT THE TOP BOX LAYOUT!!!
I finally have figured out the layout for the top boxes of my homepage. Through this I have discovered the power of flex box and also media queries for responsiveness! My main issue was wrapping my head around how to make it so rather than 3 rows, it turns into 2 when the window is smaller than a certain amount.
I figured it out with using flex-basis applied to a div which acts as a line break for the layout.
Yea I'm clever, I know : D