Updating the home page
Still working on styling the home page! My next goal is to have the playlist cards be responsive such as when the window gets bigger the cards on each row become from 2 to 4 to 6 to 9 cards on each row. My main issue is maintaining the square aspect ratio of the picture. We'll see what hack I come up with :+1: