Onto Step 2 of the Spotify Api Documentation
I am now onto Step 2 of the Spotify Api. I am now stuck on getting the access token using the authorization code that I acquired. Such as, I don't understand how to get it/ how it works. Its hard but I see a light at the end of the tunnel.