Soooooo I went back to plain vanilla JS
I decided to go back to vanilla JS rather than using a library so that I can learn more in depth of JS as a whole.

Right now I am stuck on the part of getting the access tokens. For some reason, the API call is not working so I will let the tomorrow me figure that out :pig: